perform nvm validation after git branch deployed
it’s not possible to perform and nvm:validate unless the current .nvmrc is deployed.  So we’ll wait to validate our node version until after we’ve deployed our source but before we do an npm:install.  This will allow us to do new branch deployment and new servers.